Ported initial parsing
This is the majority of code for the parser that @josephlewis42 wrote, lightly modified as I learned yacc. The major changes are a new method for parsing functions (not yet implemented), a changed method for generating transforms, and some renaming.

In the new pipescript, $, if, and switch are all implemented as normal transforms, since transforms are now much more powerful (they are given iterators, and don't have to be one to one). This also allows them to be used as busybox-style commands, since they are exposed to the registry, unlike the builtins.

Also implemented Script copy, which will be useful for implementing advanced transforms (such as switch).

This commit is the finish of the general structure of pipescript. if/switch still need to be implmented, parsing of functions also needs to be implmented, but the Parse() function is working, and everything is linked together.

Last change was prepending lowercase p (or parser) to parser commands so they are not exported.