Resources, docs, sentiment, wc
This is a pretty large change. It includes markdown documentation compiled in as resources, adds a word count and sentiment transform (which is enabled by compiled-in resource AFINN wordlist).

It is a natural inflection point in the work - the API will next undergo some breaking changes to enable simpler use